add routable macro tests
