+++ packages/yew-router-macro/Cargo.toml
@@ -14,3 +14,8 @@ heck = "0.3.2"
 proc-macro2 = "1.0.24"
 quote = "1.0.9"
 syn = { version = "1.0.64", features = ["full"] }
+
+[dev-dependencies]
+rustversion = "1.0"
+trybuild = "1.0"
+yew-router = { path = "../yew-router" }
