+++ packages/yew-router-macro/tests/routable_derive/bad-ats-fail.rs
@@ -1,0 +1,13 @@
+#[derive(yew_router::Routable)]
+enum Routes {
+    One,
+}
+
+#[derive(yew_router::Routable)]
+enum RoutesTwo {
+    #[at("/one")]
+    #[at("/two")]
+    One,
+}
+
+fn main() {}
