+++ packages/yew-router-macro/tests/routable_derive/bad-ats-fail.stderr
@@ -1,0 +1,11 @@
+error: at attribute must be present on every variant
+ --> $DIR/bad-ats-fail.rs:3:5
+  |
+3 |     One,
+  |     ^^^
+
+error: only one at attribute must be present
+ --> $DIR/bad-ats-fail.rs:8:5
+  |
+8 |     #[at("/one")]
+  |     ^
