+++ packages/yew-router-macro/tests/routable_derive/struct-fail.rs
@@ -1,0 +1,6 @@
+#[derive(yew_router::Routable)]
+struct Test {
+
+}
+
+fn main() {}
