+++ packages/yew-router-macro/tests/routable_derive/struct-fail.stderr
@@ -1,0 +1,5 @@
+error: expected enum, found struct
+ --> $DIR/struct-fail.rs:2:1
+  |
+2 | struct Test {
+  | ^^^^^^
