+++ packages/yew-router-macro/tests/routable_derive/unnamed-fields-fail.rs
@@ -1,0 +1,7 @@
+#[derive(yew_router::Routable)]
+enum Routes {
+    #[at("/one/:two")]
+    One(u32),
+}
+
+fn main() {}
