+++ packages/yew-router-macro/tests/routable_derive/unnamed-fields-fail.stderr
@@ -1,0 +1,5 @@
+error: only named fields are supported for dynamic paths
+ --> $DIR/unnamed-fields-fail.rs:3:5
+  |
+3 |     #[at("/one/:two")]
+  |     ^
