+++ packages/yew-router-macro/tests/routable_derive_test.rs
@@ -1,0 +1,9 @@
+#[allow(dead_code)]
+#[rustversion::attr(stable(1.51), test)]
+fn tests() {
+    let t = trybuild::TestCases::new();
+    t.pass("tests/routable_derive/*-pass.rs");
+    t.compile_fail("tests/routable_derive/*-fail.rs");
+}
+
+fn main() {}
